+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Removed `init` function ([#1216](https://github.com/neptune-ai/neptune-client/pull/1216))
 - Removed `get_last_run` function ([#1217](https://github.com/neptune-ai/neptune-client/pull/1217))
 - Removed `run` parameter from `init_run` function ([#1218](https://github.com/neptune-ai/neptune-client/pull/1218))
+- Monitoring is off by default for interactive Python kernels ([#1219](https://github.com/neptune-ai/neptune-client/pull/1219))
 
 ## neptune-client 0.16.18